# fse: Optimize compression