Two patches to increase compression speed. Details in the commit messages, combined effect:

```
name                     old speed      new speed      delta
Compress/gettysburg-8     181MB/s ± 1%   189MB/s ± 0%   +4.52%  (p=0.000 n=10+9)
Compress/digits-8         241MB/s ± 0%   251MB/s ± 1%   +4.09%  (p=0.000 n=9+9)
Compress/twain-8          218MB/s ± 0%   228MB/s ± 0%   +4.46%  (p=0.000 n=10+10)
Compress/low-ent-8        239MB/s ± 0%   244MB/s ± 1%   +1.89%  (p=0.000 n=10+10)
Compress/superlow-ent-8   208MB/s ± 1%   210MB/s ± 0%   +1.04%  (p=0.000 n=9+8)
Compress/endzerobits-8   11.5MB/s ± 1%  13.4MB/s ± 1%  +16.82%  (p=0.000 n=10+10)
Compress/pngdata.001-8    224MB/s ± 0%   225MB/s ± 1%   +0.79%  (p=0.000 n=8+9)
Compress/normcount2-8    35.7MB/s ± 1%  36.4MB/s ± 1%   +2.03%  (p=0.000 n=10+10)
[Geo mean]                120MB/s        125MB/s        +4.35%
```